Fix (hack) swiper loop mode
Swiper duplicates slides to support loop mode, but this duplication does not work well with nested interactive content such as as nested swipers. This fix uses empty slides initially and fills only the visible slides with the available nested swiper instances. It might not work well in the case of few slides and fast swiping.