Check sha256 sums of PDF files in cache
If the sha256 sum of a file in the cache does not match the sha256 given in the settings, the file is downloaded again. If the sha256 still doesn’t match afterwards, the nested slider for this particular snapshot is disabled. PDF pages will not be rendered and printing and mailing are disabled.