Packaging: Exclude some large files of the native hummus module
When native modules are build, intermediate files and several copies of the build result are stored in the module directory. In the case of hummus, this increases the size of the redistributable package by about 80MB. Note that I excluded these files based on implementation details of hummus 1.0.66. It may not work properly in future versions.