Fix layout problems
Swiper 4.x seems to implement auto slidesPerView differently and it seems necessary to fix the number slides per view to 3. In order to make the layout look similar/close to identical to what we had before, we need to extends the horizontal sliders width beyond the 1920px and cut off the overflow. Also, the layout algorithm for the vertical sliders seems to have changed and I needed to fix the height there as well.