Fix unwanted duplication of articles in the settings file
Upon each initialization of the main process, the settings file is merged with the default values. merge-json merges arrays if they differ and this results in duplicate entries as soon as the list of articles deviates from the defaults. Therefore, each restart causes the list of articles to grow by the number of articles in the defaults. merge-json has been removed in favor of lodash since its merge function does not merge arrays, only object properties, and therefore does not lead to duplicate articles.