Fix dupicate events in on-screen keyboard
The on-screen keyboard inserts letters based on touch and mouse events. However, in case of a touch screen, both, mouse and touch events are fire resulting in two identical letters being inserted. This commit fixes it by switching from mouse and touch events to the unified pointer events in the on-screen keyboard code.